FM-848 Move Cas3ConfirmationEntity and Cas3ConfirmationRepository to main package
