+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/cas3/jpa/entity/Cas3ConfirmationEntity.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.cas3.jpa.entity.v2
+package uk.gov.justice.digital.hmpps.approvedpremisesapi.cas3.jpa.entity
 
 import jakarta.persistence.Entity
 import jakarta.persistence.FetchType
@@ -8,7 +8,6 @@ import jakarta.persistence.OneToOne
 import jakarta.persistence.Table
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas3.jpa.entity.Cas3BookingEntity
 import java.time.OffsetDateTime
 import java.util.Objects
 import java.util.UUID
